requirements.txt
django
django-unfold
wagtail
